week8/requirements.txt
pandas
faiss-cpu
numpy
sentence-transformers
transformers
torch
